moved dead files to Archive and fixed broken test import
